fix(cuda): widen bf16<->f32 bridge length to 64-bit
import_local_bf16_as_f32 / local_f32_as_bf16 launched bf16_bits_to_f32 /
f32_to_bf16_bits with an i32 length guard (backend_cuda.rs:462,490) and int
n/idx in the kernel (bridge.cu). These bridge full-sequence tensors
([1, seq, hidden]); at seq=262144 the flat length exceeds 2^31, so the
i32::try_from guard hard-errored the 256K forward ('bf16 import length
exceeds i32') — a second index wall distinct from the GDN conv1d offset
widened in 445369d45.

Match the elementwise-kernel convention: unsigned long long n, idx =
static_cast<ull>(blockIdx.x)*blockDim.x + threadIdx.x. Chunked CE / per-row
kernels stay i32 (their flat length is bounded by window*vocab, not seq).

Found by the remote 262144 ladder run (clean GPU, 61.5 GB free — index wall,
not OOM). Bench: pending-remote, measured gate is the resumed T5 256K ladder.

+++ crates/autograd/src/backend_cuda/kernels/bridge.cu
@@ -1,9 +1,10 @@
 extern "C" __global__ void bf16_bits_to_f32(
     const unsigned short* __restrict__ input,
     float* __restrict__ output,
-    int n
+    unsigned long long n
 ) {
-    int idx = blockIdx.x * blockDim.x + threadIdx.x;
+    unsigned long long idx =
+        static_cast<unsigned long long>(blockIdx.x) * blockDim.x + threadIdx.x;
     if (idx >= n) {
         return;
     }
@@ -14,9 +15,10 @@ extern "C" __global__ void bf16_bits_to_f32(
 extern "C" __global__ void f32_to_bf16_bits(
     const float* __restrict__ input,
     unsigned short* __restrict__ output,
-    int n
+    unsigned long long n
 ) {
-    int idx = blockIdx.x * blockDim.x + threadIdx.x;
+    unsigned long long idx =
+        static_cast<unsigned long long>(blockIdx.x) * blockDim.x + threadIdx.x;
     if (idx >= n) {
         return;
     }
